feat(user): support fractional balances (#28)
Switch `User.balance` from `int` to `double` so we can represent cents
and other small currency units. This avoids rounding errors when we
apply pro-rated credits, refunds, or exchange rates across multiple
operations.

Code that assumed an `int` may fail to compile or misbehave at runtime,
so update serializers, calculations, and any validation rules that touch
`balance`. Ensure persistence and API layers accept decimals and default
to `0.0` where appropriate.

BREAKING CHANGE: The type of `User.balance` changed from `int` to
`double`. Downstream consumers must migrate to decimal handling and
review arithmetic.

+++ lib/pages/login/user_login.dart
@@ -12,7 +12,7 @@ class User {
   final String? lastName;
   final String? gender;
   final String? phoneNumber;
-  final int balance;
+  final double balance;
   final int banCount;
 
   User({
